Refactor query related structs to own module
