+++ src/missed_transfer.rs
@@ -1,10 +1,9 @@
 use std::rc::Rc;
 use tokio_core::reactor;
-use web3::futures::future;
 use web3::futures::prelude::*;
 use web3::futures::sync::mpsc;
 use web3::futures::try_ready;
-use web3::types::{Address, BlockNumber, FilterBuilder, TransactionReceipt, H256, U256};
+use web3::types::{Address, BlockNumber, FilterBuilder, U256};
 use web3::DuplexTransport;
 
 use super::contracts::TRANSFER_EVENT_SIGNATURE;
@@ -13,201 +12,6 @@ use super::transfer::Transfer;
 
 pub const LOOKBACK_RANGE: u64 = 1_000;
 pub const LOOKBACK_LEEWAY: u64 = 5;
-
-#[derive(Clone)]
-pub enum QueryChain {
-    Home,
-    Side,
-}
-
-#[derive(Clone)]
-pub struct HashQuery {
-    pub chain: QueryChain,
-    pub tx_hash: H256,
-}
-
-/// Future to handle the Stream of missed transfers by checking them, and approving them
-pub struct HandleQueries {
-    future: Box<Future<Item = (), Error = ()>>,
-}
-
-impl HandleQueries {
-    /// Returns a newly created HandleMissedTransfers Future
-    ///
-    /// # Arguments
-    ///
-    /// * `source` - Network where the missed transfers are captured
-    /// * `target` - Network where the transfer will be approved for a withdrawal
-    /// * `handle` - Handle to spawn new futures
-    pub fn new<T: DuplexTransport + 'static>(
-        homechain: &Rc<Network<T>>,
-        sidechain: &Rc<Network<T>>,
-        rx: mpsc::UnboundedReceiver<HashQuery>,
-        handle: &reactor::Handle,
-    ) -> Self {
-        let handle = handle.clone();
-        let homechain = homechain.clone();
-        let sidechain = sidechain.clone();
-        let query_future = rx.for_each(move |query| {
-            let homechain = homechain.clone();
-            let sidechain = sidechain.clone();
-            let handle = handle.clone();
-            let (source, target) = match query.chain {
-                QueryChain::Home => (homechain, sidechain),
-                QueryChain::Side => (sidechain, homechain),
-            };
-            FindTransferInTransaction::new(&source, &query.tx_hash)
-                .and_then(move |transfers| {
-                    let handle = handle.clone();
-                    let target = target.clone();
-                    let futures: Vec<ValidateAndApproveTransfer<T>> = transfers
-                        .iter()
-                        .map(move |transfer| {
-                            let handle = handle.clone();
-                            let target = target.clone();
-                            ValidateAndApproveTransfer::new(&target, &handle, &transfer)
-                        })
-                        .collect();
-                    future::join_all(futures)
-                })
-                .and_then(|_| Ok(()))
-                .map_err(move |e| {
-                    error!("error approving queried transfers: {:?}", e);
-                })
-        });
-        HandleQueries {
-            future: Box::new(query_future),
-        }
-    }
-}
-
-impl Future for HandleQueries {
-    type Item = ();
-    type Error = ();
-    fn poll(&mut self) -> Poll<Self::Item, Self::Error> {
-        self.future.poll()
-    }
-}
-
-pub enum FindTransferState {
-    ExtractTransfers(Box<Future<Item = Vec<Transfer>, Error = ()>>),
-    FetchReceipt(Box<Future<Item = Option<TransactionReceipt>, Error = ()>>),
-}
-
-/// Future to find a vec of transfers at a specific transaction
-pub struct FindTransferInTransaction<T: DuplexTransport + 'static> {
-    hash: H256,
-    source: Rc<Network<T>>,
-    state: FindTransferState,
-}
-
-impl<T: DuplexTransport + 'static> FindTransferInTransaction<T> {
-    /// Returns a newly created FindTransferInTransaction Future
-    ///
-    /// # Arguments
-    ///
-    /// * `source` - Network where the transaction took place
-    /// * `hash` - Transaction hash to check
-    fn new(source: &Rc<Network<T>>, hash: &H256) -> Self {
-        let web3 = source.web3.clone();
-        let future = web3.clone().eth().transaction_receipt(*hash).map_err(|e| {
-            error!("error getting transaction receipt: {:?}", e);
-        });
-        let state = FindTransferState::FetchReceipt(Box::new(future));
-        FindTransferInTransaction {
-            hash: *hash,
-            source: source.clone(),
-            state,
-        }
-    }
-}
-
-impl<T: DuplexTransport + 'static> Future for FindTransferInTransaction<T> {
-    type Item = Vec<Transfer>;
-    type Error = ();
-    fn poll(&mut self) -> Poll<Self::Item, Self::Error> {
-        loop {
-            let source = self.source.clone();
-            let hash = self.hash;
-            let next = match self.state {
-                FindTransferState::ExtractTransfers(ref mut future) => {
-                    let transfers = try_ready!(future.poll());
-                    if transfers.is_empty() {
-                        warn!("no relay transactions found at {}", self.hash);
-                        return Err(());
-                    } else {
-                        return Ok(Async::Ready(transfers));
-                    }
-                }
-                FindTransferState::FetchReceipt(ref mut future) => {
-                    let receipt = try_ready!(future.poll());
-                    match receipt {
-                        Some(r) => {
-                            if r.block_hash.is_none() {
-                                error!("receipt did not have block hash");
-                                return Err(());
-                            }
-                            let block_hash = r.block_hash.unwrap();
-                            r.block_number.map_or_else(
-                                || {
-                                    error!("receipt did not have block number");
-                                    Err(())
-                                },
-                                |receipt_block| {
-                                    let future = source
-                                        .web3
-                                        .clone()
-                                        .eth()
-                                        .block_number()
-                                        .and_then(move |block| {
-                                            let mut transfers = Vec::new();
-                                            if let Some(confirmed) =
-                                                block.checked_rem(receipt_block).map(|u| u.as_u64())
-                                            {
-                                                if confirmed > source.confirmations {
-                                                    let logs = r.logs;
-                                                    for log in logs {
-                                                        if log.topics[0] == TRANSFER_EVENT_SIGNATURE.into()
-                                                            && log.topics[3] == source.relay.address().into()
-                                                        {
-                                                            let destination: Address = log.topics[1].into();
-                                                            let amount: U256 = log.data.0[..32].into();
-                                                            if destination == Address::zero() {
-                                                                info!("found mint. Skipping");
-                                                                continue;
-                                                            }
-                                                            let transfer = Transfer {
-                                                                destination,
-                                                                amount,
-                                                                tx_hash: hash,
-                                                                block_hash,
-                                                                block_number: receipt_block,
-                                                            };
-                                                            transfers.push(transfer);
-                                                        }
-                                                    }
-                                                }
-                                            }
-                                            Ok(transfers)
-                                        })
-                                        .map_err(|e| {
-                                            error!("error getting current block: {:?}", e);
-                                        });
-                                    Ok(FindTransferState::ExtractTransfers(Box::new(future)))
-                                },
-                            )
-                        }
-                        None => {
-                            error!("Could not find requested transaction hash");
-                            return Err(());
-                        }
-                    }?
-                }
-            };
-            self.state = next;
-        }
-    }
-}
 
 /// Stream of Transfer that were missed, either from downtime, or failed approvals
 pub struct FindMissedTransfers(mpsc::UnboundedReceiver<Transfer>);
